refactor: align Runtime picker component placement
