+++ src/handlers/runtime/list/screen.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router";
 import type { ScreenProps } from "../../types";
-import { RuntimePicker } from "../components/RuntimePicker";
+import { RuntimePicker } from "../../../components/RuntimePicker";
 
 export function RuntimeListScreen(props: ScreenProps) {
   const navigate = useNavigate();
